Изменены комментарии, убраны ненужные файлы
